reorganize claim utils
