+++ src/claim/claim-utils.test.js
@@ -5,6 +5,7 @@ import { Entry } from './entry';
 const chai = require('chai');
 const snarkjs = require('snarkjs');
 const { babyJub } = require('circomlib');
+const claimUtils = require('./claim-utils');
 const claim = require('./claim');
 const utils = require('../utils');
 
@@ -21,7 +22,7 @@ describe('[claim-utils]', () => {
                      + '0000000000000000000000000000000000004d59000000010000000000000004';
     const entry = Entry.newFromHex(entryHex);
     // increase version
-    claim.incClaimVersion(entry);
+    claimUtils.incClaimVersion(entry);
     const claimExample = claim.newClaimFromEntry(entry);
     expect(claimExample).to.be.not.equal(undefined);
     if (claimExample == null) { return; }
@@ -31,12 +32,12 @@ describe('[claim-utils]', () => {
 
   it('Clear most significant byte', () => {
     const testIn = '0xf6f36d94c84a7096c572b83d44df576e1ffb3573123f62099f8d4fa19de806bd';
-    const testOutBuff = claim.clearElemMostSignificantByte(utils.hexToBytes(testIn));
+    const testOutBuff = claimUtils.clearElemMostSignificantByte(utils.hexToBytes(testIn));
     const testInSim = '0x00f36d94c84a7096c572b83d44df576e1ffb3573123f62099f8d4fa19de806bd';
     expect(testInSim).to.be.equal(utils.bytesToHex(testOutBuff));
 
     const testIn2 = '0x00f36d94c84a7096c572b83d44df576e1ffb3573123f62099f8d4fa19de806bd';
-    const testOutBuff2 = claim.clearElemMostSignificantByte(utils.hexToBytes(testIn2));
+    const testOutBuff2 = claimUtils.clearElemMostSignificantByte(utils.hexToBytes(testIn2));
     expect(testIn2).to.be.equal(utils.bytesToHex(testOutBuff2));
   });
 
@@ -44,17 +45,17 @@ describe('[claim-utils]', () => {
     // Check hash is more than 32 bytes
     const hash0 = '0xf6a4f36d94c84a7096c572b83d44df576e1ffb3573123f62099f8d4fa19de806bd';
     expect(() => {
-      claim.checkElemFitsClaim(utils.hexToBytes(hash0));
+      claimUtils.checkElemFitsClaim(utils.hexToBytes(hash0));
     }).to.throw('Element is not 32 bytes length');
     // Check hash is not valid for claim field
     const hash1 = '0xf6f36d94c84a7096c572b83d44df576e1ffb3573123f62099f8d4fa19de806bd';
     expect(() => {
-      claim.checkElemFitsClaim(utils.hexToBytes(hash1));
+      claimUtils.checkElemFitsClaim(utils.hexToBytes(hash1));
     }).to.throw('Element does not fit on claim element size');
     // Check hash is valid
     const hash2 = '0x06f36d94c84a7096c572b83d44df576e1ffb3573123f62099f8d4fa19de806bd';
     expect(() => {
-      claim.checkElemFitsClaim(utils.hexToBytes(hash2));
+      claimUtils.checkElemFitsClaim(utils.hexToBytes(hash2));
     }).not.to.throw();
   });
 
@@ -63,10 +64,10 @@ describe('[claim-utils]', () => {
     const test = babyJub.p;
     const test1 = test.add(bigInt(1));
     expect(() => {
-      claim.checkElemFitsClaim(utils.bigIntToBufferBE(test));
+      claimUtils.checkElemFitsClaim(utils.bigIntToBufferBE(test));
     }).not.to.throw();
     expect(() => {
-      claim.checkElemFitsClaim(utils.bigIntToBufferBE(test1));
+      claimUtils.checkElemFitsClaim(utils.bigIntToBufferBE(test1));
     }).to.throw('Element does not fit on claim element size');
   });
 });
