+++ src/crypto/crypto.js
@@ -1,5 +1,7 @@
 const utilsBabyJub = require('./babyjub-utils');
+const eddsaBabyJub = require('./eddsa-babyjub');
 
 module.exports = {
   utilsBabyJub,
+  eddsaBabyJub,
 };
